asset: move group key specific code into new file group_key.go
